refactor: created map-engine.js and began cleanup of app.js
