+++ package.json
@@ -10,8 +10,7 @@
     "lodash": "^4.14.2",
     "prompt": "^1.0.0"
   },
-  "devDependencies": {
-  },
+  "devDependencies": {},
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
     "serve": "node src/app.js"
